+++ examples/complete/variables.tf
@@ -12,3 +12,9 @@ variable "tags" {
     LayerId            = "cExample"
   }
 }
+
+variable "prefix" {
+  type        = string
+  description = "Prefix for the resource name"
+  default     = "external"
+}