+++ CHANGELOG.md
@@ -8,7 +8,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 - Feature: DocDB Global cluster.
 - fix CKV2_AWS_5: "Ensure that Security Groups are attached to another resource"
-- showcase usage of the following: cluster_identifier_prefix, external db_cluster_parameter_group_name, snapshot_identifier and promotion_tier
+
+## [1.3.2] - 2023-11-23
+- fix: usage of cluster_identifier_prefix
+- fix: usage of external db_cluster_parameter_group_name
+- fix: usage of promotion_tier
+- fix: usage of cluster snapshot and snapshot_identifier
 
 ## [1.3.1] - 2023-11-16
 - fix: usage of apply_immediately
@@ -49,7 +54,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.0.1] - 2022-03-22
 - Feature: Document db cluster Instance
 - Feature: DocumentDB SubnetGroup
-- Feature/fix
 
 ## [1.0.0] - YYYY-MM-DD
 ### Description
@@ -57,8 +61,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added: DocumentDB example
 - Feature/fix
 
-[Unreleased]: https://github.com/boldlink/terraform-aws-docdb/compare/1.3.1...HEAD
+[Unreleased]: https://github.com/boldlink/terraform-aws-docdb/compare/1.3.2...HEAD
 
+[1.3.2]: https://github.com/boldlink/terraform-aws-docdb/releases/tag/1.3.2
 [1.3.1]: https://github.com/boldlink/terraform-aws-docdb/releases/tag/1.3.1
 [1.3.0]: https://github.com/boldlink/terraform-aws-docdb/releases/tag/1.3.0
 [1.2.1]: https://github.com/boldlink/terraform-aws-docdb/releases/tag/1.2.1